Network.NBD

Documentation

nBD_DEFAULT_PORT :: Int

The default (IANA-assigned) port for NBD

data NbdExportFlag

Flags of a single export

Constructors

HasFlags
Should always be set

ReadOnly
The export is read-only

SendFlush
The server supports NBD_CMD_FLUSH commands

SendFua
The server supports the NBD_CMD_FLAG_FUA flag

Rotational
The client should schedule I/O accesses as for a rotational medium

SendTrim
The server supports NBD_CMD_TRIM commands

data NbdCommandFlag

Flags set on NBD commands

Constructors

ForceUnitAccess
Force Unit Access on NBD_CMD_WRITE

type ExportName = Text

A type synonym for export names In line with the protocol spec, this will be UTF-8 encoded

type ByteCount = CSize

type FileOffset = COff

data ProtocolException

The ProtocolException type lists a couple of values which can be raised at runtime

Constructors

InvalidClientFlags !Word32
Client sent invalid flags during negotiation

InvalidMagic String !Word64
An invalid magic value was received

ParseFailure String
Parsing of some field failed

ClientAbort
The client sent an NBD_OPT_ABORT option

InvalidProtocolHeader ByteString
Server sent some invalid/unknown protocol header

InvalidServerFlags !Word16
Server sent invalid flags during negotiation TODO Not very happy with this one, shouldn't pass a Word32 as-is but decode to NbdReply

OptionError !Word32
Server returned an error reply to an option command

UnexpectedServerReply String

data Handle

Opaque type representing a command handle sent by a client