Network.NBD.Client

Documentation

negotiateNewstyle :: MonadIO m => GLConduit ByteString m ByteString

Run a newstyle negotiation handshake with the server

optionAbort :: Monad m => GSource m ByteString

Send the Abort option to the server

optionList :: MonadIO m => forall u. Pipe ByteString ByteString ByteString u m (Maybe [ExportName])

Send the List option to the server. If the server returns a list of export names, this will be returned. If this is disabled (by policy), Nothing is returned.

optionExportName :: MonadIO m => ExportName -> Pipe ByteString ByteString ByteString u m (ByteCount, [NbdExportFlag])

Send the ExportName option to the server

read :: Monad m => Handle -> [NbdCommandFlag] -> FileOffset -> ByteCount -> GSource m ByteString

Send a Read command to the server

write :: Monad m => Handle -> [NbdCommandFlag] -> FileOffset -> ByteCount -> ByteString -> GSource m ByteString

Send a Write command to the server

disconnect :: Monad m => Handle -> [NbdCommandFlag] -> GSource m ByteString

Send a Disconnect command to the server

flush :: Monad m => Handle -> [NbdCommandFlag] -> GSource m ByteString

Send a Flush command to the server

trim :: Monad m => Handle -> [NbdCommandFlag] -> FileOffset -> ByteCount -> GSource m ByteString

Send a Trim command to the server

getResponse :: MonadIO m => (Handle -> Maybe ByteCount) -> GLSink ByteString m Response

Retrieve a single Response from the server. The passed function should return the expected payload size (in case of a Read request), or Nothing, based on the handle.

data Response

Server response to a client command

Constructors

Success Handle
Command succeeded

Error Handle Errno
Command execution failed with given error

Data Handle ByteString
Command succeeded and returned some data

handle0 :: Handle

An initial Handle

newHandle :: Handle -> Handle

Calculate a new Handle given a used one