Network.NBD.Server

Documentation

negotiateNewstyle

Arguments

:: MonadIO m

=> Maybe [ExportName]
An optional list of export names. If Nothing, export listing support will be disabled.

-> Pipe ByteString ByteString ByteString () m ExportName

Run the new-style protocol negotiation handshake with a client The result will be the name of the export the client wants to connect to.

sendExportInformation

Arguments

:: Monad m

=> ByteCount
Size of the selected export

-> [NbdExportFlag]
Flags set for the export

-> Pipe l i ByteString u m ()

After negotiation, send information about the selected export

getCommand :: MonadIO m => Pipe ByteString ByteString o u m Command

Receive a single command from the client

sendReply :: Monad m => Handle -> Pipe l i ByteString u m ()

Send a reply to the client indicating success for the given command handle

sendReplyData :: Monad m => Handle -> ByteString -> Pipe l i ByteString u m ()

Send a reply to the client indicating success for the given command handle. The provided data will be passed along (e.g. in response to a Read command).

sendReplyData' :: Monad m => Handle -> (ByteString -> m ()) -> m ()

Construct a reply to the client indicating success for the given handle. The reply bytes will be passed to the given action, which is responsible of pushing this data to the client, maybe followed by more data (e.g. when replying to a Read command).

sendError :: Monad m => Handle -> Errno -> Pipe l i ByteString u m ()

Send an error reply to the client for the given handle

data Command

Representation of a client command

Constructors

Read

Fields

readHandle :: !Handle

Request handle

readFrom :: !FileOffset

Requested offset

readLength :: !ByteCount

Requested length

readFlags :: [NbdCommandFlag]

Command flags

Write

Fields

writeHandle :: !Handle

Request handle

writeFrom :: !FileOffset

Write offset

writeData :: ByteString

Data

writeFlags :: [NbdCommandFlag]

Command flags

Disconnect

Fields

disconnectFlags :: [NbdCommandFlag]

Command flags

Flush

Fields

flushHandle :: !Handle

flushFlags :: [NbdCommandFlag]

Command flags

Trim

Fields

trimHandle :: !Handle

Request handle

trimFrom :: !FileOffset

Requested offset

trimLength :: !ByteCount

Requested length

trimFlags :: [NbdCommandFlag]

Command flags

UnknownCommand

Fields

unknownCommandId :: !Word32

Request command number

unknownCommandHandle :: !Handle

Request handle

unknownCommandOffset :: !FileOffset

Request offset

unknownCommandLength :: !ByteCount

Request length

unknownCommandFlags :: [NbdCommandFlag]

Command flags