Network.NBD.Server.Simple

Action result representations

data Response a

Success indicator returned by action handlers

Constructors

OK a
Success

Error Errno
Failure. The given Errno will be returned to the client

data ReadResponse

Response data returned by a ReadHandler

Constructors

Data ByteString
Return given data as request reply

Sendfile Fd FileRange
Use sendfile to return some data

data FileRange

File range for sendfile.

Constructors

EntireFile

PartOfFile

Fields

rangeOffset :: Integer

rangeLength :: Integer

Action handlers

data ExportHandler

Action vtable to handle client requests. Any Maybe field can be set to Nothing, which implies the given feature is not available on the server. The feature flags sent to a client during protocol negotiation will be based on this.

Constructors

ExportHandler

Fields

handleRead :: ReadHandler

Handle a Read request

handleWrite :: Maybe WriteHandler

Handle a Write request

handleFlush :: Maybe FlushHandler

Handle a Flush request

handleTrim :: Maybe TrimHandler

Handle a Trim request

handleDisconnect :: Maybe DisconnectHandler

Action to run prior to disconnecting a client due to a Disconnect request

handleUnknownCommand :: Maybe UnknownCommandHandler

Handle an unknown request

handleSize :: ByteCount

Size of the export

handleHasFUA :: Bool

Server supports the ForceUnitAccess request flag

handleIsRotational :: Bool

Server exposes a rotational device (which might impact the client IO scheduler)

type FileOffset = COff

type ByteCount = CSize

data NbdCommandFlag

Flags set on NBD commands

Constructors

ForceUnitAccess
Force Unit Access on NBD_CMD_WRITE

type Flags = [NbdCommandFlag]

Type synonym for a list of NbdCommandFlags

type ReadHandler = FileOffset -> ByteCount -> Flags -> IO (Response ReadResponse)

Read request handler prototype

type WriteHandler = FileOffset -> ByteString -> Flags -> IO (Response ())

Write request handler prototype

type FlushHandler = Flags -> IO (Response ())

Flush request handler prototype

type TrimHandler = FileOffset -> ByteCount -> Flags -> IO (Response ())

Trim request handler prototype

type DisconnectHandler = Flags -> IO ()

Disconnect request handler prototype. Note the connection will be closed by the server when receiving a Disconnect request, after executing any given handler, even if the handler throws an Exception.

type CommandId = Word32

Command identifier for requests received from a client

type UnknownCommandHandler = CommandId -> FileOffset -> ByteCount -> Flags -> IO (Response ())

Unknown request handler prototype

Server execution

type ExportMaker = SockAddr -> ExportName -> IO (Maybe ExportHandler)

Prototype of an action turning a SockAddr and ExportName into an export, or rather its ExportHandler vtable. If Nothing is returned, this implies the given ExportName was invalid, and the server should act accordingly.

data ServerSettings

Constructors

TCPServerSettings Int HostPreference

UNIXServerSettings FilePath

data HostPreference

Which host to bind.

Note: The IsString instance recognizes the following special values:

* means HostAny

*4 means HostIPv4

*6 means HostIPv6

Constructors

HostAny

HostIPv4

HostIPv6

Host String

nBD_DEFAULT_PORT :: Int

The default (IANA-assigned) port for NBD

runServer

Arguments

:: ServerSettings
Listen address

-> Maybe [ExportName]
Optional list of export names to expose to the client. If Nothing, export listing is disabled.

-> ExportMaker
Action to retrieve an ExportHandler for a given client connection

-> IO ()

Run an NBD server

Utilities

withBoundsCheck

Arguments

:: Monad m

=> ByteCount
Size of the export we're dealing with

-> FileOffset
Offset of the request

-> ByteCount
Length of the request

-> m (Response a)
Action to execute if request is within bounds

-> m (Response a)

Utility action to validate request boundaries. This action will return eINVAL to the client if an out-of-bounds request was received