Network.NBD

Documentation

nBD_DEFAULT_PORT :: Int

The default (IANA-assigned) port for NBD

data NbdExportFlag

Flags of a single export

Constructors

HasFlags
Should always be set

ReadOnly
The export is read-only

SendFlush
The server supports NBD_CMD_FLUSH commands

SendFua
The server supports the NBD_CMD_FLAG_FUA flag

Rotational
The client should schedule I/O accesses as for a rotational medium

SendTrim
The server supports NBD_CMD_TRIM commands

data NbdCommandFlag

Flags set on NBD commands

Constructors

ForceUnitAccess
Force Unit Access on NBD_CMD_WRITE

type ExportName = Text

A type synonym for export names In line with the protocol spec, this will be UTF-8 encoded

type ExportSize = Word64

A type synonym for the size of a single export

type Offset = Word64

A type synonym for offset values passed in commands

type Length = Word32

A type synonym for length values passed in commands

data ProtocolException

The ProtocolException type lists a couple of values which can be raised at runtime

Constructors

InvalidClientFlags !Word32
Client sent invalid flags during negotiation

InvalidMagic String !Word64
An invalid magic value was received

ParseFailure String
Parsing of some field failed

ClientAbort
The client sent an NBD_OPT_ABORT option